Add files via upload
Attached is the current version of the demo script to analyze slap2 test data. It's built from code that Alex Maier wrote and extended from there. 
The code is still messy and quite hacky in places, but I wanted to share it with you in case anyone wants to have a look and build on it.